[move-decorators] Move widget story decorators to respective widget directories
Right now, all the widget decorators are in the testutils directory.

It might make more sense to have them in their direct widget storybook directories,
localizing them to where they're actually used.

Issue: none

Test plan:
`pnpm lint` `pnpm typecheck`

Confirm no changes in UI Tests check below.
