+++ packages/perseus/src/widgets/image/__docs__/image-renderer-decorator.tsx
@@ -8,7 +8,7 @@ import {
 } from "@khanacademy/perseus-core";
 import * as React from "react";
 
-import QuestionRendererForStories from "./question-renderer-for-stories";
+import QuestionRendererForStories from "../../__testutils__/question-renderer-for-stories";
 
 export const imageRendererDecorator = (_, {args, parameters}) => {
     return (
