+++ packages/perseus/src/widgets/interactive-graphs/__docs__/interactive-graph-renderer-decorator.tsx
@@ -1,9 +1,9 @@
 import {generateInteractiveGraphQuestion} from "@khanacademy/perseus-core";
 import * as React from "react";
 
-import QuestionRendererForStories from "./question-renderer-for-stories";
+import QuestionRendererForStories from "../../__testutils__/question-renderer-for-stories";
 
-import type {APIOptions} from "../../types";
+import type {APIOptions} from "../../../types";
 import type {
     PerseusInteractiveGraphWidgetOptions,
     PerseusRenderer,
