+++ packages/perseus/src/widgets/label-image/__docs__/label-image-initial-state-regression.stories.tsx
@@ -1,5 +1,6 @@
 import {themeModes} from "../../../../../../.storybook/modes";
-import {labelImageRendererDecorator} from "../../__testutils__/label-image-renderer-decorator";
+
+import {labelImageRendererDecorator} from "./label-image-renderer-decorator";
 
 import type {PerseusLabelImageWidgetOptions} from "@khanacademy/perseus-core";
 import type {Meta, StoryObj} from "@storybook/react-vite";
