+++ packages/perseus/src/widgets/label-image/__docs__/label-image-renderer-decorator.tsx
@@ -6,7 +6,7 @@ import {
 } from "@khanacademy/perseus-core";
 import * as React from "react";
 
-import {ServerItemRendererWithDebugUI} from "../../testing/server-item-renderer-with-debug-ui";
+import {ServerItemRendererWithDebugUI} from "../../../testing/server-item-renderer-with-debug-ui";
 
 export const labelImageRendererDecorator = (_, {args, parameters}) => {
     return (
