+++ packages/perseus/src/widgets/matcher/__docs__/matcher-renderer-decorator.tsx
@@ -4,7 +4,7 @@ import {
 } from "@khanacademy/perseus-core";
 import * as React from "react";
 
-import QuestionRendererForStories from "./question-renderer-for-stories";
+import QuestionRendererForStories from "../../__testutils__/question-renderer-for-stories";
 
 export const matcherRendererDecorator = (_, {args}) => {
     return (
