+++ packages/perseus/src/widgets/matrix/__docs__/matrix-initial-state-regression.stories.tsx
@@ -1,6 +1,7 @@
 import {themeModes} from "../../../../../../.storybook/modes";
-import {matrixRendererDecorator} from "../../__testutils__/matrix-renderer-decorator";
 import {rtlDecorator} from "../../__testutils__/story-decorators";
+
+import {matrixRendererDecorator} from "./matrix-renderer-decorator";
 
 import type {MatrixPublicWidgetOptions} from "@khanacademy/perseus-core";
 import type {Meta, StoryObj} from "@storybook/react-vite";
