+++ packages/perseus/src/widgets/numeric-input/__docs__/numeric-input-renderer-decorator.tsx
@@ -1,17 +1,16 @@
 import {
-    generateMatrixOptions,
-    generateMatrixWidget,
+    generateNumericInputOptions,
+    generateNumericInputWidget,
     generateTestPerseusRenderer,
 } from "@khanacademy/perseus-core";
 import * as React from "react";
 
-import QuestionRendererForStories from "./question-renderer-for-stories";
+import QuestionRendererForStories from "../../__testutils__/question-renderer-for-stories";
 
-import type {APIOptions} from "../../types";
+import type {APIOptions} from "../../../types";
 import type {UserInputMap} from "@khanacademy/perseus-core";
-import type {Decorator} from "@storybook/react-vite";
 
-export const matrixRendererDecorator: Decorator = (
+export const numericInputRendererDecorator = (
     _,
     {
         args,
@@ -28,10 +27,12 @@ export const matrixRendererDecorator: Decorator = (
     return (
         <QuestionRendererForStories
             question={generateTestPerseusRenderer({
-                content: parameters?.content ?? "[[☃ matrix 1]]",
+                content:
+                    parameters?.content ??
+                    "Registry numbers for USS Enterprise: [[☃ numeric-input 1]]",
                 widgets: {
-                    "matrix 1": generateMatrixWidget({
-                        options: generateMatrixOptions({
+                    "numeric-input 1": generateNumericInputWidget({
+                        options: generateNumericInputOptions({
                             ...args,
                         }),
                     }),
