+++ packages/perseus/src/widgets/radio/__docs__/radio-renderer-decorator.tsx
@@ -6,11 +6,11 @@ import {
 } from "@khanacademy/perseus-core";
 import * as React from "react";
 
-import {ServerItemRenderer} from "../../server-item-renderer";
-import {ServerItemRendererWithDebugUI} from "../../testing/server-item-renderer-with-debug-ui";
-import {testDependenciesV2} from "../../testing/test-dependencies";
+import {ServerItemRenderer} from "../../../server-item-renderer";
+import {ServerItemRendererWithDebugUI} from "../../../testing/server-item-renderer-with-debug-ui";
+import {testDependenciesV2} from "../../../testing/test-dependencies";
 
-import type {APIOptions} from "../../types";
+import type {APIOptions} from "../../../types";
 import type {
     PerseusImageDetail,
     RadioDefaultWidgetOptions,
